+++ samples/twitter-bot/src/lib.rs
@@ -4,14 +4,14 @@ use distri::{
     servers::registry::{register_mcp_servers, McpServerRegistry, ServerMetadata, ServerTrait},
     types::{Configuration, TransportType},
 };
+use std::sync::Arc;
+
+use distri::{agent::ExecutorContext, types::McpSession, ToolSessionStore};
 
 use distri_server::agent_server::DistriAgentServer;
 use dotenv::dotenv;
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::RwLock;
-
-use crate::store::get_tools_session_store;
-mod store;
 
 pub fn get_agent_server() -> DistriAgentServer {
     DistriAgentServer {
@@ -74,4 +74,31 @@ pub async fn init_agent_executor(config: &Configuration) -> anyhow::Result<Arc<A
             .await?;
     }
     Ok(executor)
+}
+
+pub struct StaticToolSessionStore {
+    session_key: String,
+}
+
+#[async_trait::async_trait]
+impl ToolSessionStore for StaticToolSessionStore {
+    async fn get_session(
+        &self,
+        _tool_name: &str,
+        _context: &ExecutorContext,
+    ) -> anyhow::Result<Option<McpSession>> {
+        Ok(Some(McpSession {
+            token: self.session_key.clone(),
+            expiry: None,
+        }))
+    }
+}
+
+pub fn get_tools_session_store() -> Arc<Box<dyn ToolSessionStore>> {
+    dotenv::dotenv().ok();
+    let session_key =
+        std::env::var("X_USER_SESSION").unwrap_or_else(|_| "test_session_key".to_string());
+    // Create executor with static session store
+
+    Arc::new(Box::new(StaticToolSessionStore { session_key }))
 }
